requirements.txt
dill=0.3.8
numpy=1.26.4
pandas=2.2.1
pillow=10.2.0
scikit-learn=1.4.1.post1
scipy=1.12.0
streamlit=1.33.0
